% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{as.matrix.textmodel_word2vec}
\alias{as.matrix.textmodel_word2vec}
\title{Extract word or document vectors}
\usage{
\method{as.matrix}{textmodel_word2vec}(x, normalize = TRUE, layer = "words", ...)
}
\arguments{
\item{x}{a \code{textmodel_wordvector} or \code{textmodel_docvector} object.}

\item{normalize}{if \code{TRUE}, returns normalized vectors.}

\item{layer}{the layer from which the vectors are extracted.}

\item{...}{not used.}
}
\value{
a matrix that contain the word vectors in rows.
}
\description{
Extract word or document vectors from a \code{textmodel_wordvector} or \code{textmodel_docvector} object.
}
