% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.doc2vec.R
\name{as.textmodel_doc2vec}
\alias{as.textmodel_doc2vec}
\title{Create a doc2vec model}
\usage{
as.textmodel_doc2vec(x, model, compound = TRUE, group_data = FALSE, ...)
}
\arguments{
\item{x}{a \link[quanteda:tokens]{quanteda::tokens} or \link[quanteda:dfm]{quanteda::dfm} object.}

\item{model}{a textmodel_wordvector object.}

\item{compound}{if \code{TRUE}, compound multi-word expressions in \code{x} based on \code{model}
internally. Only applies when \code{x} is a \link[quanteda:tokens]{quanteda::tokens} object.}

\item{group_data}{if \code{TRUE}, apply \code{dfm_group(x)} before creating document vectors.}

\item{...}{additional arguments passed to the underlying function.}
}
\value{
Returns a textmodel_doc2vec object with the following elements:
\item{values}{a list of matrices for word and document vectors.}
\item{dim}{the size of the document vectors.}
\item{concatenator}{the concatenator in \code{x}.}
\item{docvars}{document variables copied from \code{x}.}
\item{call}{the command used to execute the function.}
\item{version}{the version of the wordvector package.}
}
\description{
Create a doc2vec model as weighted word vectors.
}
\details{
For Japanese or Chinese texts, \code{model$concatenator} must be empty ("").
The value is inherited from the tokens object on which the model was trained.
It triggers tokenization of words in \code{model} and compounding of characters in \code{x}
before creating document vectors.
}
