% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_word2vec}
\alias{is_word2vec}
\alias{is_doc2vec}
\alias{check_word2vec}
\alias{check_doc2vec}
\alias{check_model}
\title{Functions to check input objects}
\usage{
is_word2vec(x)

is_doc2vec(x)

check_word2vec(x)

check_doc2vec(x)

check_model(x, allow = c("word2vec", "doc2vec", "lsa"))
}
\arguments{
\item{x}{input object.}

\item{allow}{type of object allowed.}
}
\description{
Functions to check input objects
}
\keyword{internal}
