% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_corpus_news2014}
\alias{data_corpus_news2014}
\title{Yahoo News summaries from 2014}
\source{
\url{https://www.yahoo.com/news/}
}
\description{
A corpus object containing 2,000 news summaries collected from Yahoo News via
RSS feeds in 2014.
}
\references{
Watanabe, K. (2018). Newsmap: A semi-supervised approach to
geographical news classification. Digital Journalism, 6(3), 294–309.
https://doi.org/10.1080/21670811.2017.1293487
}
