% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsa.R
\name{lsa}
\alias{lsa}
\title{Latent Semantic Analysis model}
\usage{
lsa(
  x,
  dim = 50,
  min_count = 5L,
  engine = c("RSpectra", "irlba", "rsvd"),
  weight = "count",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a \link[quanteda:tokens]{quanteda::tokens} object.}

\item{dim}{the size of the word vectors.}

\item{min_count}{the minimum frequency of the words. Words less frequent than this in the \code{tokens} object are removed before training.}

\item{engine}{select the engine perform SVD to generate word vectors.}

\item{weight}{weighting scheme passed to \code{\link[quanteda:dfm_weight]{quanteda::dfm_weight()}}.}

\item{verbose}{if \code{TRUE}, print the progress of training.}

\item{...}{additional arguments.}
}
\value{
Returns a fitted textmodel_wordvector with the following elements:
\item{vectors}{a matrix for word vectors.}
\item{frequency}{the frequency of words in \code{x}.}
\item{engine}{the SVD engine used.}
\item{weight}{weighting scheme.}
\item{concatenator}{the concatenator in \code{x}.}
\item{call}{the command used to execute the function.}
\item{version}{the version of the wordvector package.}
}
\description{
Train a Latent Semantic Analysis model (Deerwester et al., 1990) on a \link[quanteda:tokens]{quanteda::tokens} object.
}
\references{
Deerwester, S. C., Dumais, S. T., Landauer, T. K., Furnas, G. W., & Harshman, R. A. (1990).
Indexing by latent semantic analysis. JASIS, 41(6), 391–407.
}
