% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{print.textmodel_doc2vec}
\alias{print.textmodel_doc2vec}
\title{Print method for trained document vectors}
\usage{
\method{print}{textmodel_doc2vec}(x, ...)
}
\arguments{
\item{x}{for print method, the object to be printed}

\item{...}{unused}
}
\value{
an invisible copy of \code{x}.
}
\description{
Print method for trained document vectors
}
\keyword{internal}
