% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{print.textmodel_docvector}
\alias{print.textmodel_docvector}
\title{Print method for trained document vectors}
\usage{
\method{print}{textmodel_docvector}(x, ...)
}
\arguments{
\item{x}{for print method, the object to be printed}

\item{...}{unused}
}
\description{
Print method for trained document vectors
}
\keyword{internal}
