% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{print.textmodel_word2vec}
\alias{print.textmodel_word2vec}
\title{Print method for trained word vectors}
\usage{
\method{print}{textmodel_word2vec}(x, ...)
}
\arguments{
\item{x}{for print method, the object to be printed}

\item{...}{not used.}
}
\value{
an invisible copy of \code{x}.
}
\description{
Print method for trained word vectors
}
\keyword{internal}
