% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probability}
\alias{probability}
\title{Generic function to extract probabilities}
\usage{
probability(x, ...)
}
\arguments{
\item{x}{an object from which probabilities are extracted.}

\item{...}{passed to underlying functions.}
}
\description{
Generic function to extract probabilities
}
