% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2vec.R
\name{textmodel_doc2vec}
\alias{textmodel_doc2vec}
\title{Create distributed representation of documents}
\usage{
textmodel_doc2vec(x, model = NULL, group_data = FALSE, ...)
}
\arguments{
\item{x}{a \link[quanteda:tokens]{quanteda::tokens} object.}

\item{model}{a textmodel_wordvector object.}

\item{group_data}{if \code{TRUE}, apply \code{dfm_group(x)} before creating document vectors.}

\item{...}{passed to \verb{[textmodel_word2vec]} when \code{model = NULL}.}
}
\value{
Returns a textmodel_docvector object with elements inherited from \code{model}
or passed via \code{...} plus:
\item{values}{a matrix for document vectors.}
\item{call}{the command used to execute the function.}
}
\description{
Create distributed representation of documents as weighted word vectors.
}
