% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weights}
\alias{weights}
\title{[experimental] Extract word vector weights}
\usage{
weights(x, mode = c("words", "values"))
}
\arguments{
\item{x}{a \code{textmodel_wordvector} object.}

\item{mode}{specify the type of resulting object.}
}
\value{
a \code{matrix} of word vector weights when \code{mode = "value"} or of
words sorted in descending order by the weights when \code{mode = "word"}.
}
\description{
[experimental] Extract word vector weights
}
