% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analogy}
\alias{analogy}
\title{Convert formula to named character vector}
\usage{
analogy(formula)
}
\arguments{
\item{formula}{a \link{formula} object that defines the relationship between words
using \code{+} or \code{-} operators.}
}
\description{
Convert formula to named character vector
}
\examples{
analogy(~ berlin - germany + france)
analogy(~ quick - quickly + slowly)
}
