% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analogy}
\alias{analogy}
\title{Convert formula to named character vector}
\usage{
analogy(formula)
}
\arguments{
\item{formula}{a \link{formula} object that defines the relationship between words
using \code{+} or \code{-} operators.}
}
\value{
a named character vector to be passed to \code{\link[=similarity]{similarity()}}.
}
\description{
Convert a formula to a named character vector in analogy tasks.
}
\examples{
analogy(~ berlin - germany + france)
analogy(~ quick - quickly + slowly)
}
\seealso{
\code{\link[=similarity]{similarity()}}
}
