% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.word2vec.R
\name{as.textmodel_word2vec}
\alias{as.textmodel_word2vec}
\title{Create a word2vec model}
\usage{
as.textmodel_word2vec(x, ...)
}
\arguments{
\item{x}{a matrix containing word vectors in rows.}

\item{...}{additional arguments.}
}
\value{
Returns a dummy textmodel_word2vec object
}
\description{
Create a word2vec model from a matrix.
}
