% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2vec.R
\name{doc2vec}
\alias{doc2vec}
\title{Create distributed representation of documents}
\usage{
doc2vec(x, model = NULL, ...)
}
\arguments{
\item{x}{a \link[quanteda:tokens]{quanteda::tokens} object.}

\item{model}{a textmodel_wordvector object.}

\item{...}{passed to \verb{[word2vec]} when \code{model = NULL}.}
}
\value{
Returns a textmodel_docvector object with elements inherited from \code{model}
or passed via \code{...} plus:
\item{vectors}{a matrix for document vectors.}
\item{call}{the command used to execute the function.}
}
\description{
Create distributed representation of documents
}
