% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perplexity}
\alias{perplexity}
\title{Compute perplexity of a model}
\usage{
perplexity(x, targets, data)
}
\arguments{
\item{x}{a trained \code{textmodel_wordvector} object.}

\item{targets}{words for which probabilities are computed.}

\item{data}{the probabilities words are tested against occurrences of words in \code{data};
should be a \link[quanteda:tokens]{quanteda::tokens} or \link[quanteda:dfm]{quanteda::dfm}.}
}
\description{
Compute the perplexity of a trained word2vec model with data.
}
\keyword{internal}
