% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{probability}
\alias{probability}
\title{Compute probability of words}
\usage{
probability(x, words, mode = c("words", "values"))
}
\arguments{
\item{x}{a \code{textmodel_wordvector} object fitted with \code{normalize = FALSE}.}

\item{words}{words for which probability is computed.}

\item{mode}{specify the type of resulting object.}
}
\value{
a \code{matrix} of probability scores when \code{mode = "values"} or of words
sorted in descending order by the probability scores when \code{mode = "words"}.
When \code{words} is a named numeric vector, probability scores are weighted by
the  values.
}
\description{
Compute the probability of words given other words.
}
\seealso{
\code{\link[=similarity]{similarity()}}
}
