% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity}
\alias{similarity}
\title{Generic function to extract similarities}
\usage{
similarity(x, ...)
}
\arguments{
\item{x}{an object from which similarities are extracted.}

\item{...}{passed to underlying functions.}
}
\description{
Generic function to extract similarities
}
