% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity}
\alias{similarity}
\title{Compute similarity between word vectors}
\usage{
similarity(x, words, mode = c("character", "numeric"))
}
\arguments{
\item{x}{a \code{textmodel_wordvector} object.}

\item{words}{words for which similarity is computed.}

\item{mode}{specify the type of resulting object.}
}
\value{
a \code{matrix} of cosine similarity scores when \code{mode = "values"} or of
words sorted in descending order by the similarity scores when \code{mode = "words"}.
When \code{words} is a named numeric vector, word vectors are weighted and summed
before computing similarity scores.
}
\description{
Compute the cosine similarity between word vectors for selected words.
}
\seealso{
\code{\link[=probability]{probability()}}
}
