% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{similarity}
\alias{similarity}
\title{Compute similarity between word vectors}
\usage{
similarity(x, words, mode = c("simil", "word"))
}
\arguments{
\item{x}{a \code{textmodel_wordvector} object.}

\item{words}{words for which similarity is computed.}

\item{mode}{specify the type of resulting object.}
}
\value{
a \code{matrix} of cosine similarity scores when \code{mode = "simil"} or of
words sorted by the similarity scores when \verb{mode = "word}.
}
\description{
Compute similarity between word vectors
}
